feat: send user code with runExampleCode api and display the result
Build Value Display Area to display testCases value and stdout value. Build ResultsArea to display Run example result, which can be runtime error or a list of results of each example cases. Now the user can write code in playground and click run to the results of each example cases.